+++ src/components/common/crew-list/crew-card.tsx
@@ -27,9 +27,7 @@ export default function CrewCard({
   inWhere,
 }: CrewCardProps) {
   const [prefetched, setPrefetched] = useState(new Set());
-  // NOTE: api연결 후 되돌리기
-  // const CREWPAGE = `/crew/detail/${id}`;
-  const CREWPAGE = `/crew/detail/1`;
+  const CREWPAGE = `/crew/detail/${id}`;
   const router = useRouter();
 
   const handleCardClick = () => {
